admin: split out transaction endpoints
